`RunInterfate.Retry`: purge data to truncate the run
